+++ Tests/WindowsCloudPC.Tests.ps1
@@ -3,6 +3,7 @@ BeforeDiscovery {
         'Connect-CloudPC',
         'Get-CloudPC',
         'Get-CloudPCByProvisioningPolicy',
+        'Get-CloudPCLaunchDetail',
         'Get-CloudPCProvisioningPolicy',
         'Get-CloudPCRemoteActionResult',
         'Get-CloudPCUsage',